Added EAE.java and Tile.java. Also, modified Player.java.
EAE.java is the main area to run the game. This is where we will combine
all the parts to make the whole. Currently still working on the whole
section, but I mainly want to get the basics of the game finished in
here, then spread off into the rest of the project.

Tile.java is the class file for the tiles that players will land on.
This is still being modified, as well.

Player.java was modifed to incorperate the player's location with the
current map loadout in EAE.java. I modified the main constructor, as the
default location of the player should be the starting point of the map,
so I removed the input "int loc" to reflect this. Also, added the
movePlayer method to incorporate the rollDice method in EAE.java, so it
moves the player to the correct spot in the map.

Everything is just sort of basic, but I plan to implement more later.